+++ sdks/outpost-python/src/outpost_sdk/models/destinationcreateazureservicebus.py
@@ -44,7 +44,7 @@ class DestinationCreateAzureServiceBusTypedDict(TypedDict):
 
     """
     delivery_metadata: NotRequired[Nullable[Dict[str, str]]]
-    r"""Static key-value pairs merged into event metadata on every delivery."""
+    r"""Static key-value pairs merged into event metadata on every attempt."""
     metadata: NotRequired[Nullable[Dict[str, str]]]
     r"""Arbitrary contextual information stored with the destination."""
 
@@ -74,7 +74,7 @@ class DestinationCreateAzureServiceBus(BaseModel):
     """
 
     delivery_metadata: OptionalNullable[Dict[str, str]] = UNSET
-    r"""Static key-value pairs merged into event metadata on every delivery."""
+    r"""Static key-value pairs merged into event metadata on every attempt."""
 
     metadata: OptionalNullable[Dict[str, str]] = UNSET
     r"""Arbitrary contextual information stored with the destination."""