+++ sdks/outpost-python/src/outpost_sdk/models/destinationcreaterabbitmq.py
@@ -41,7 +41,7 @@ class DestinationCreateRabbitMQTypedDict(TypedDict):
 
     """
     delivery_metadata: NotRequired[Nullable[Dict[str, str]]]
-    r"""Static key-value pairs merged into event metadata on every delivery."""
+    r"""Static key-value pairs merged into event metadata on every attempt."""
     metadata: NotRequired[Nullable[Dict[str, str]]]
     r"""Arbitrary contextual information stored with the destination."""
 
@@ -71,7 +71,7 @@ class DestinationCreateRabbitMQ(BaseModel):
     """
 
     delivery_metadata: OptionalNullable[Dict[str, str]] = UNSET
-    r"""Static key-value pairs merged into event metadata on every delivery."""
+    r"""Static key-value pairs merged into event metadata on every attempt."""
 
     metadata: OptionalNullable[Dict[str, str]] = UNSET
     r"""Arbitrary contextual information stored with the destination."""