+++ sdks/outpost-python/src/outpost_sdk/models/destinationhookdeck.py
@@ -47,7 +47,7 @@ class DestinationHookdeckTypedDict(TypedDict):
     r"""ISO Date when the destination was last updated."""
     config: NotRequired[Any]
     delivery_metadata: NotRequired[Nullable[Dict[str, str]]]
-    r"""Static key-value pairs merged into event metadata on every delivery."""
+    r"""Static key-value pairs merged into event metadata on every attempt."""
     metadata: NotRequired[Nullable[Dict[str, str]]]
     r"""Arbitrary contextual information stored with the destination."""
     target: NotRequired[str]
@@ -90,7 +90,7 @@ class DestinationHookdeck(BaseModel):
     config: Optional[Any] = None
 
     delivery_metadata: OptionalNullable[Dict[str, str]] = UNSET
-    r"""Static key-value pairs merged into event metadata on every delivery."""
+    r"""Static key-value pairs merged into event metadata on every attempt."""
 
     metadata: OptionalNullable[Dict[str, str]] = UNSET
     r"""Arbitrary contextual information stored with the destination."""