+++ sdks/outpost-python/src/outpost_sdk/models/destinationupdategcppubsub.py
@@ -30,7 +30,7 @@ class DestinationUpdateGCPPubSubTypedDict(TypedDict):
     config: NotRequired[GCPPubSubConfigTypedDict]
     credentials: NotRequired[GCPPubSubCredentialsTypedDict]
     delivery_metadata: NotRequired[Nullable[Dict[str, str]]]
-    r"""Static key-value pairs merged into event metadata on every delivery."""
+    r"""Static key-value pairs merged into event metadata on every attempt."""
     metadata: NotRequired[Nullable[Dict[str, str]]]
     r"""Arbitrary contextual information stored with the destination."""
 
@@ -54,7 +54,7 @@ class DestinationUpdateGCPPubSub(BaseModel):
     credentials: Optional[GCPPubSubCredentials] = None
 
     delivery_metadata: OptionalNullable[Dict[str, str]] = UNSET
-    r"""Static key-value pairs merged into event metadata on every delivery."""
+    r"""Static key-value pairs merged into event metadata on every attempt."""
 
     metadata: OptionalNullable[Dict[str, str]] = UNSET
     r"""Arbitrary contextual information stored with the destination."""